CET_Prediction/cet-web-app/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.5.2
xgboost==2.1.3
openpyxl==3.1.5
joblib==1.4.2
gunicorn==21.2.0
python-dotenv==1.0.1
google-generativeai
python-dotenv
flask-cors